unit_tests: In TestDataMethods, added testCongestionMark. Refs #4449.

+++ tests/src/net/named_data/jndn/tests/unit_tests/TestDataMethods.java
@@ -718,6 +718,6 @@ public class TestDataMethods {
     data.wireDecode(element, TlvWireFormat.get());
     data.setLpPacket(lpPacket);
 
-    assertEquals(1, data.getCongestionMark());
+    assertEquals((long)1, data.getCongestionMark());
   }
 }
